Update Prefix-Cache-Scorer `cache_tracking` Mode with The v0.2 KVCache.Indexer (#245)
* Support for new version of kv-cache-manager is added to scheduler

* Redundant function parameter and struct member are removed

* Minor changes in comments

* Redundant setting of ZMQEndpoint is removed

* - general refactoring
- enhanced prefix-cache-scorer cache-tracking mode's configurability
- updated docs
- enhanced docs (no HTML use)


* refactor example configs


---------
