+++ deploy/config/epp-prefix-cache-tracking-config.yaml
@@ -3,20 +3,28 @@
 apiVersion: inference.networking.x-k8s.io/v1alpha1
 kind: EndpointPickerConfig
 plugins:
-- type: single-profile-handler
-- type: decode-filter
-- type: prefix-cache-scorer
-  parameters:
-    mode: cache_tracking
-    kvCacheRedisAddr: ${REDIS_HOST}:${REDIS_PORT}
-- type: load-aware-scorer
-- type: max-score-picker
+  - type: single-profile-handler
+  - type: decode-filter
+  - type: prefix-cache-scorer
+    parameters:
+      mode: cache_tracking
+      indexerConfig:
+        tokenProcessorConfig:
+          blockSize: 64                 # must match vLLM block size
+          hashSeed: "42"                # must match vLLM PYTHONHASHSEED env var
+        kvBlockIndexConfig:
+          enableMetrics: true           # enable kv-block index metrics (prometheus)
+  - type: kv-cache-scorer # kv-cache-utilization
+  - type: queue-scorer
+  - type: max-score-picker
 schedulingProfiles:
-- name: default
-  plugins:
-  - pluginRef: decode-filter
-  - pluginRef: prefix-cache-scorer
-    weight: 2.0
-  - pluginRef: load-aware-scorer
-    weight: 1.0
-  - pluginRef: max-score-picker
+  - name: default
+    plugins:
+      - pluginRef: decode-filter
+      - pluginRef: prefix-cache-scorer
+        weight: 3.0
+      - pluginRef: kv-cache-scorer
+        weight: 1.0
+      - pluginRef: queue-scorer
+        weight: 1.0
+      - pluginRef: max-score-picker
